improved the service worker response for cached assets
if the cached asset is older than the file on the server, the service worker now caches the new asset instead.
This is done seperately from the response to the current request in order to minimise reaction time; the next time a user asks for it, he'll get the new asset.